user verification added user.questions component

+++ app/Livewire/User/QuestionsComponent.php
@@ -3,12 +3,15 @@
 namespace App\Livewire\User;
 
 use App\Models\Question;
+use App\Models\Topic;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Component;
 
 class QuestionsComponent extends Component
 {
     public $question;
     public $title;
+    public $topicUserId;
     public function render()
     {
         return view('livewire.user.questions-component');
@@ -17,22 +20,33 @@ class QuestionsComponent extends Component
         $validated = $this->validate([
             'title' => 'required',
         ]);
-        $this->question->title = $validated['title'];
-        $this->question->save();
+        if ($this->topicUserId == Auth::user()->id) {
+            $this->question->title = $validated['title'];
+            $this->question->save();
+        }
+        
     }
     public function mount(){
         $this->title = $this->question->title;
+        $this->topicUserId = Topic::firstWhere('id', $this->question->topic_id)->user_id;
+
     }
 
     public function deleteQuestion(): void{
         
         //$this->topic->();
         $id = $this->question->id;
-        Question::destroy($this->question->id);
-        $this->dispatch("topicDelete", id: $id);
+        if ($this->topicUserId == Auth::user()->id) {
+            Question::destroy($this->question->id);
+            $this->dispatch("topicDelete", id: $id);
+        }
+        
     }
 
     public function activateQuestion(): void{        
-        $this->dispatch("activateQuestion", id: $this->question->id);
+        
+        if ($this->topicUserId == Auth::user()->id) {
+            $this->dispatch("activateQuestion", id: $this->question->id);
+        }
     }
 }
